Major login logout ui updates bbk
