+++ ios/Runner/Info.plist
@@ -41,5 +41,12 @@
 	</array>
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
+	<key>NSLocationWhenInUseUsageDescription</key>
+  	<string>Reason why app needs location</string>
+  	<key>NSLocationAlwaysAndWhenInUseUsageDescription</key>
+  	<string>Reason why app needs location</string>
+  
+  	<key>NSLocationAlwaysUsageDescription</key>
+  	<string>Reason why app needs location</string>
 </dict>
 </plist>
