+++ lib/pages/home/login/userprofile.dart
@@ -1,0 +1,145 @@
+import 'package:flutter/material.dart';
+import 'dart:convert';
+import 'dart:io';
+import 'package:touristguide/pages/afterloginhome.dart';
+
+// import 'pages/profile.dart';
+// import './pages/home/home.dart';
+// import 'package:fluttertoast/fluttertoast.dart';
+// import 'package:touristguide/pages/home/login/login_page.dart';
+// import 'package:material_search/material_search.dart';
+// import 'package:touristguide/pages/home/search/search.dart';
+import 'package:http/http.dart' as http;
+
+class UserProfile extends StatefulWidget{
+  final String uname;
+
+  UserProfile({Key key, this.uname}) : super(key: key);
+
+  UserProfileState createState()=>  UserProfileState();
+}
+
+
+class UserProfileState extends State<UserProfile> {
+  static String tag = 'home-page';
+  var _tst;
+  String datas, tid;
+
+   @override
+  void initState() {
+    super.initState();
+    tid = "${widget.uname}";
+    print(tid);
+    datas = tid;
+    _fetchData(tid);
+  }
+
+  _fetchData(String tid) async {
+    String username = 'beingbivek@gmail.com';
+    String password = 'bivek';
+    print(tid);
+    String basicAuth =
+        'Basic ' + base64Encode(utf8.encode('$username:$password'));
+    //print(basicAuth);
+
+    final url = "http://192.168.100.4:8090/tourists/$tid";
+    try {
+      final response = await http.get(url, headers: {
+        HttpHeaders.AUTHORIZATION: basicAuth,
+        HttpHeaders.CONTENT_TYPE: 'application/json'
+      });
+
+      if (response.statusCode == 200) {
+        final responseJson = json.decode(response.body);
+        //final user = responseJson['id'];
+
+        setState(() {
+          this._tst = responseJson;
+        });
+        print(_tst);
+        //print(responseJson);
+      } else {
+        // _tst = 'Error getting response:\nHttp status ${response.statusCode}';
+      }
+
+      //print(map["List"]);
+
+    } catch (exception) {
+      setState(() {
+        // this._tst = "Failed parsing response because of: $exception";
+      });
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final alucard = Hero(
+      tag: 'hero',
+      child: Padding(
+        padding: EdgeInsets.all(16.0),
+        child: CircleAvatar(
+          radius: 72.0,
+          backgroundColor: Colors.transparent,
+          backgroundImage: AssetImage('assets/alucard.jpg'),
+        ),
+      ),
+    );
+
+    final welcome = Padding(
+      padding: EdgeInsets.all(8.0),
+      child:
+       Text(_tst['name'],
+        style: TextStyle(fontSize: 28.0, color: Colors.white),
+      ),
+    );
+
+    final lorem = Padding(
+      padding: EdgeInsets.all(8.0),
+      child: Text(
+        'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec hendrerit condimentum mauris id tempor. Praesent eu commodo lacus. Praesent eget mi sed libero eleifend tempor. Sed at fringilla ipsum. Duis malesuada feugiat urna vitae convallis. Aliquam eu libero arcu.',
+        style: TextStyle(fontSize: 16.0, color: Colors.white),
+      ),
+    );
+
+    final homebutton = Padding(
+      padding: EdgeInsets.symmetric(vertical: 16.0),
+      child: Material(
+        borderRadius: BorderRadius.circular(30.0),
+        shadowColor: Colors.lightBlueAccent.shade200,
+        elevation: 5.0,
+        child: MaterialButton(
+          minWidth: 200.0,
+          height: 42.0,
+          onPressed: () {
+            AfterLoginHome();
+            var route = new MaterialPageRoute(
+              builder: (BuildContext context) =>
+                  new AfterLoginHome(pvalue: datas),
+            );
+            Navigator.of(context).push(route);
+          },
+          color: Colors.lightBlueAccent,
+          child: Text('Home', style: TextStyle(color: Colors.white)),
+        ),
+      ),
+    );
+
+    final body = Container(
+      width: MediaQuery.of(context).size.width,
+      padding: EdgeInsets.all(28.0),
+      decoration: BoxDecoration(
+        gradient: LinearGradient(colors: [
+          Colors.blue,
+          Colors.lightBlueAccent,
+        ]),
+      ),
+      child: Column(
+        children: <Widget>[alucard, welcome, lorem, homebutton],
+      ),
+    );
+
+    return Scaffold(
+      body: body,
+    );
+  }
+}
