+++ pubspec.yaml
@@ -37,7 +37,6 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   assets:
-    - assets/logo.png
     - assets/alucard.jpg
   fonts:
       - family: Nunito
